refactor: rename repositories->repository
