+++ internal/repository/storage.go
@@ -1,4 +1,4 @@
-package repositories
+package repository
 
 type Gauge float64
 type Counter int64
