+++ internal/service/memstorage.go
@@ -1,64 +1,63 @@
 package service
 
 import (
+	"github.com/Hobrus/hobrusmetrics.git/internal/repository"
 	"sync"
-
-	"github.com/Hobrus/hobrusmetrics.git/internal/repositories"
 )
 
 type MemStorage struct {
 	mu       sync.RWMutex
-	gauges   map[string]repositories.Gauge
-	counters map[string]repositories.Counter
+	gauges   map[string]repository.Gauge
+	counters map[string]repository.Counter
 }
 
 func NewMemStorage() *MemStorage {
 	return &MemStorage{
-		gauges:   make(map[string]repositories.Gauge),
-		counters: make(map[string]repositories.Counter),
+		gauges:   make(map[string]repository.Gauge),
+		counters: make(map[string]repository.Counter),
 	}
 }
 
-func (m *MemStorage) UpdateGauge(name string, value repositories.Gauge) {
+func (m *MemStorage) UpdateGauge(name string, value repository.Gauge) {
 	m.mu.Lock()
 	defer m.mu.Unlock()
 	m.gauges[name] = value
 }
 
-func (m *MemStorage) UpdateCounter(name string, value repositories.Counter) {
+func (m *MemStorage) UpdateCounter(name string, value repository.Counter) {
 	m.mu.Lock()
 	defer m.mu.Unlock()
 	m.counters[name] += value
 }
 
-func (m *MemStorage) GetGauge(name string) (repositories.Gauge, bool) {
+func (m *MemStorage) GetGauge(name string) (repository.Gauge, bool) {
 	m.mu.RLock()
 	defer m.mu.RUnlock()
 	value, exists := m.gauges[name]
 	return value, exists
 }
 
-func (m *MemStorage) GetCounter(name string) (repositories.Counter, bool) {
+func (m *MemStorage) GetCounter(name string) (repository.Counter, bool) {
 	m.mu.RLock()
 	defer m.mu.RUnlock()
 	value, exists := m.counters[name]
 	return value, exists
 }
 
-func (m *MemStorage) GetAllGauges() map[string]repositories.Gauge {
+func (m *MemStorage) GetAllGauges() map[string]repository.Gauge {
 	m.mu.RLock()
 	defer m.mu.RUnlock()
-	gauges := make(map[string]repositories.Gauge)
+	gauges := make(map[string]repository.Gauge)
 	for k, v := range m.gauges {
 		gauges[k] = v
 	}
 	return gauges
 }
 
-func (m *MemStorage) GetAllCounters() map[string]repositories.Counter {
+func (m *MemStorage) GetAllCounters() map[string]repository.Counter {
 	m.mu.RLock()
 	defer m.mu.RUnlock()
-	counters := make(map[string]repositories.Counter)
+	counters := make(map[string]repository.Counter)
 	for k, v := range m.counters {
 		counters[k] = v
 	}
